rename flash decoding kernel
